# packages/stock-data/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GlobalStockDataService",
    "description": "面向股票分析、量化交易与策略回测的全球股票数据服务 (50GB空间极致优化架构)",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/snapshot": {
      "get": {
        "tags": [
          "Real-time Batch Market Snapshot"
        ],
        "summary": "Get Snapshots Get",
        "description": "【批量实时股票行情快照】GET 方式：\n支持跨市场 (A股/港股/美股/ETF/指数) 一次性获取多只股票的最新快照 (最新价、涨跌幅、成交量额、最高最低等)。\n单次最多支持 200 只股票。\n严格遵循真实性原则: 若标的无实时行情或代码无效，记录于 missing 列表，严禁伪造假数据。",
        "operationId": "get_snapshots_get_api_v1_snapshot_get",
        "parameters": [
          {
            "name": "symbols",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "逗号分隔的标的代码列表，如 600519.SH.STK,AAPL.US.STK,0700.HK.STK,510300.SH.ETF",
              "title": "Symbols"
            },
            "description": "逗号分隔的标的代码列表，如 600519.SH.STK,AAPL.US.STK,0700.HK.STK,510300.SH.ETF"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SnapshotBatchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/snapshot/batch": {
      "post": {
        "tags": [
          "Real-time Batch Market Snapshot"
        ],
        "summary": "Get Snapshots Batch",
        "description": "【批量实时股票行情快照】POST 方式：\n接收 JSON 请求体 {\"symbols\": [\"600519.SH.STK\", \"AAPL.US.STK\", ...]}，\n适合自选股列表、持仓看板、组合跟踪等大批量标的一次性拉取。\n单次最多支持 200 只股票。\n严格遵循真实性原则: 若标的无实时行情或代码无效，记录于 missing 列表，严禁伪造假数据。",
        "operationId": "get_snapshots_batch_api_v1_snapshot_batch_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SnapshotBatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SnapshotBatchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SnapshotBatchRequest": {
        "properties": {
          "symbols": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Symbols",
            "description": "批量股票代码列表，如 ['600519.SH.STK', 'AAPL.US.STK']"
          }
        },
        "type": "object",
        "required": [
          "symbols"
        ],
        "title": "SnapshotBatchRequest"
      },
      "SnapshotBatchResponse": {
        "properties": {
          "count": {
            "type": "integer",
            "title": "Count",
            "description": "获取到的有效实时快照数量"
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/SnapshotItem"
            },
            "type": "array",
            "title": "Data",
            "description": "有效行情快照列表"
          },
          "missing": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Missing",
            "description": "无实时行情或代码未识别的标的列表 (严格拒绝虚构数据)"
          }
        },
        "type": "object",
        "required": [
          "count"
        ],
        "title": "SnapshotBatchResponse"
      },
      "SnapshotItem": {
        "properties": {
          "symbol": {
            "type": "string",
            "title": "Symbol",
            "description": "规范化标的代码: [TICKER].[MARKET].[TYPE]"
          },
          "ticker": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ticker",
            "description": "标的代码/简码 (如 600519, AAPL)"
          },
          "name": {
            "type": "string",
            "title": "Name",
            "description": "标的名称"
          },
          "latest_price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Price",
            "description": "最新成交价格 (无实时成交时为 None)"
          },
          "pre_close": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pre Close",
            "description": "昨收价"
          },
          "open": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Open",
            "description": "今开盘价"
          },
          "high": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "High",
            "description": "今日最高价"
          },
          "low": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Low",
            "description": "今日最低价"
          },
          "change": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Change",
            "description": "涨跌额"
          },
          "pct_change": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pct Change",
            "description": "涨跌幅(%)"
          },
          "volume": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Volume",
            "description": "今日成交量"
          },
          "amount": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amount",
            "description": "今日成交额"
          },
          "turnover_rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Turnover Rate",
            "description": "换手率(%)"
          },
          "pe": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pe",
            "description": "市盈率(动态/TTM)"
          },
          "pe_ttm": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pe Ttm",
            "description": "市盈率(TTM)"
          },
          "pb": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pb",
            "description": "市净率"
          },
          "total_market_cap": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Market Cap",
            "description": "总市值 (亿元/亿美元/亿港元)"
          },
          "market_cap": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Market Cap",
            "description": "总市值 (亿元/亿美元/亿港元)"
          },
          "circulating_market_cap": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Circulating Market Cap",
            "description": "流通市值 (亿元/亿美元/亿港元)"
          },
          "float_market_cap": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Float Market Cap",
            "description": "流通市值 (亿元/亿美元/亿港元)"
          },
          "dividend_yield": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dividend Yield",
            "description": "股息率(%)"
          },
          "dividend_yield_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dividend Yield Pct",
            "description": "股息率(%)"
          },
          "nav": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nav",
            "description": "ETF 实时单位净值/IOPV (仅 ETF 标的有效)"
          },
          "premium_rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Premium Rate",
            "description": "ETF 折溢价率(%) = (最新价 - NAV) / NAV * 100"
          },
          "ask_prices": {
            "anyOf": [
              {
                "items": {
                  "type": "number"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ask Prices",
            "description": "卖一至卖五档挂单价格 (仅 A股支持，其他市场为 None)"
          },
          "ask_volumes": {
            "anyOf": [
              {
                "items": {
                  "type": "number"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ask Volumes",
            "description": "卖一至卖五档挂单量 (手)"
          },
          "bid_prices": {
            "anyOf": [
              {
                "items": {
                  "type": "number"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bid Prices",
            "description": "买一至买五档挂单价格 (仅 A股支持，其他市场为 None)"
          },
          "bid_volumes": {
            "anyOf": [
              {
                "items": {
                  "type": "number"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bid Volumes",
            "description": "买一至买五档挂单量 (手)"
          },
          "timestamp": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timestamp",
            "description": "行情时间戳 (UTC 毫秒)"
          }
        },
        "type": "object",
        "required": [
          "symbol",
          "name"
        ],
        "title": "SnapshotItem"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}